refactor: #202 SortableCategoryRow 컴포넌트 분리 및 계층 구조 드래그 지원
